+++ next.config.mjs
@@ -1,4 +1,21 @@
 /** @type {import('next').NextConfig} */
-const nextConfig = {};
+const nextConfig = {
+    reactStrictMode: true,
+    typescript: {
+        ignoreBuildErrors: true
+    },
+    images: {
+        domains: ['*']
+    },
+    webpack(config) {
+        config.module.rules.push({
+            test: /\.svg$/i,
+            issuer: /\.[jt]sx?$/,
+            use: ['@svgr/webpack']
+        });
+        // return the modified config
+        return config;
+    },
+};
 
 export default nextConfig;
