+++ src/modules/wall/submit.tsx
@@ -7,6 +7,11 @@ import { ModalComponent } from '@/components/Modal';
 const Submit = () => {
     const [openModal, setOpenModal] = useState<boolean>(false);
 
+    // submit post
+    const onSubmitPost = async () => {
+        console.log('hee hee')
+    }
+
     return (
         <div>
             {/* button */}
@@ -20,6 +25,7 @@ const Submit = () => {
             <ModalComponent
                 openModal={openModal}
                 onClose={() => setOpenModal(false)}
+                onClick={() => { onSubmitPost(); }}
                 title="Hit me with your best thought!"
                 buttonAction="Submit"
             />
